requirements.txt
Flask>=2.0
flask-cors>=6.0.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
Flask-Limiter>=3.0 
gunicorn>=20.0
redis>=4.0